requirements.txt
branca==0.5.0
feather_format==0.4.1
folium==0.12.1
geojson==3.1.0
geopandas==0.10.2
pandas==1.4.3
Shapely==1.8.2
